requirements.txt
feedparser~=6.0.11
newspaper3k
pandas~=2.3.1
scikit-learn
spacy~=3.8.7
pyyaml~=6.0.2
schedule~=1.2.2
requests~=2.32.4
sqlalchemy~=2.0.41
spacy
streamlit~=1.47.1

beautifulsoup4~=4.13.4